+++ anybox/recipe/odoo/tests/test_server.py
@@ -395,6 +395,32 @@ class TestServer(RecipeTestCase):
             expected.append('test_openerp')
         self.assertScripts(expected)
 
+    def test_install_scripts_80_no_devtools(self):
+        self.test_install_scripts_80(with_devtools=False)
+
+    def test_gunicorn_preload_databases(self, databases='onedb',
+                                        expected="('onedb',)"):
+        self.make_recipe(version='local %s' % os.path.join(TEST_DIR, 'odoo80'),
+                         gunicorn='direct')
+        self.recipe.version_detected = "7.0"
+        self.recipe.options['gunicorn.preload_databases'] = databases
+        self.recipe.options['options.log_handler'] = ":INFO,werkzeug:WARNING"
+
+        self.install_scripts()
+
+        self.assertScripts(['gunicorn_openerp'])
+        gunicorn_conf = os.path.join(self.recipe.etc,
+                                     'gunicorn_openerp.conf.py')
+        self.assertTrue(os.path.exists(gunicorn_conf))
+        with open(gunicorn_conf) as conf:
+            for line in conf:
+                if 'preload_dbs =' in line:
+                    self.assertEqual(line.strip(), "preload_dbs = " + expected)
+
+    def test_gunicorn_preload_databases_multiple(self):
+        self.test_gunicorn_preload_databases(databases='db1\ndb2',
+                                             expected="('db1', 'db2')")
+
     def test_install_scripts_80_server_wide_modules(self):
         self.make_recipe(version='local %s' % os.path.join(TEST_DIR, 'odoo80'),
                          gunicorn='direct',